Fix Windows compilation

+++ src/neural_network/layer/MaxPooling2D.cpp
@@ -36,7 +36,7 @@ std::vector<float> MaxPooling2D::computeOutput(const std::vector<float>& inputs,
 {
     auto output = vector<float>(this->numberOfOutputs, numeric_limits<float>::lowest());
     const int rest = this->shapeOfInput[0] % sizeOfFilterMatrix == 0 ? 0 : 1;
-    for (size_t i = 0; i < inputs.size(); ++i)
+    for (int i = 0; i < (int)inputs.size(); ++i)
     {
         const int indexOutputX = (i % this->shapeOfInput[0]) / this->sizeOfFilterMatrix;
         const int indexOutputY = (i / this->shapeOfInput[0]) / this->sizeOfFilterMatrix;
